Eliminate Sonar code smell
Because having more than 7 parameters will cause people's heads to explode due to brain overload, instead of suppressing this error, I decided to refactor this to use the builder directly in the constructor and eliminate the 8 parameter constructor. This should shut up sonar and keep dev's fragile heads from exploding. :P

We may want to do this in the other model builders for consistency, but I am not going to do that with this PR. The other constructors are under or at the 7 parameter limit.