EDGPATRON-69: Update to Log4J 2.16.0 (CVE-2021-4104).
The referenced ticket is originally for CVE-2021-44228.
However, that CVE does not exactly affect the 1.2 versions of Log4J (per-say).
A separate CVE (CVE-2021-4104) appears to have been created to reference a way to exploit the JNDI expoit mentioned in CVE-2021-4428 via the JMSAppender.

This required several additional code changes and should be reviewed for breaking changes.
I have tested that, with the changes applied, the logs do work with 2.16.0.
I am not familiar with `jcl-over-slf4j` and as such I am concerned with potential for breakage there.

The dependencies now look like:
[INFO] -----------------------< org.folio:edge-patron >------------------------
[INFO] Building Edge API - Patron Empowerment 4.5.0
[INFO] --------------------------------[ jar ]---------------------------------
[INFO]
[INFO] --- maven-dependency-plugin:2.8:tree (default-cli) @ edge-patron ---
[INFO] org.folio:edge-patron:jar:4.5.0
[INFO] +- org.folio:edge-common:jar:2.0.2:compile
[INFO] |  \- log4j:log4j:jar:1.2.17:compile
[INFO] +- org.apache.logging.log4j:log4j-api:jar:2.16.0:compile
[INFO] +- org.apache.logging.log4j:log4j-core:jar:2.16.0:compile
[INFO] \- org.apache.logging.log4j:log4j-1.2-api:jar:2.16.0:compile
[INFO] ------------------------------------------------------------------------

Using an exclusion on log4j:log4j in org.folio:edge-common causes logs to not show up.
This concerns me that more work is needed.