EDGPATRON-96: Suppress headers, edge-common 4.4.0, fix timeout, enable compression
edge-patron passes on all headers from the external client to mod-patron.

This includes the Host: header. This fails when edge-patron and mod-patron run a different hosts.

edge-patron should never pass any headers for security reasons unless strictly required.

To fix this problem changes in edge-common are needed, therefore this is blocked by the release of edge-common 4.4.0.